rename(fraud): TheoricalDistanceAndDurationCheck
